Move sequence map/set to platform core.
