+++ platform-sdk/swirlds-common/src/main/java/module-info.java
@@ -44,9 +44,6 @@ module com.swirlds.common {
     exports com.swirlds.common.metrics.platform.prometheus;
     exports com.swirlds.common.notification;
     exports com.swirlds.common.platform;
-    exports com.swirlds.common.sequence;
-    exports com.swirlds.common.sequence.map;
-    exports com.swirlds.common.sequence.set;
     exports com.swirlds.common.stream;
     exports com.swirlds.common.stream.internal;
     exports com.swirlds.common.threading;
