+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/sequence/set/SequenceSet.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.swirlds.common.sequence.set;
+package com.swirlds.platform.sequence.set;
 
-import com.swirlds.common.sequence.Shiftable;
 import com.swirlds.common.utility.Clearable;
+import com.swirlds.platform.sequence.Shiftable;
 import java.util.List;
 import java.util.function.Consumer;
